requirements.txt
# fastapi
# uvicorn[standard]
# jinja2
# httpx                  # For async HTTP requests (optional)
# requests               # For web scraping
# beautifulsoup4
# ibm-watsonx-ai         # Your IBM Watsonx SDK
# google-genai   # For Gemini API
# python-multipart       # If you want form-data (optional)

# fastapi
# uvicorn[standard]
# jinja2
# httpx
# requests
# beautifulsoup4
# ibm-watsonx-ai
# google-genai
# python-multipart
# python-dotenv


fastapi
uvicorn[standard]
jinja2
httpx
requests
beautifulsoup4
ibm-watsonx-ai
google-genai
python-multipart
